fixed problems
Swapped names of row and chars in Row Methods for both encryption and
decryption functions.

Swapped out existing charsInRow-decryption Method for a new one.

New charsInRows decryption Method is a modified version of the
row-decryption Method. Added a few console.log and test calls for
testing in node.

Also changed Math.ceil in charsInRow en- and decrypt to Math.floor.